requirements.txt
pytesseract
pdf2image
transformers
torch
pillow
numpy
opencv-python
scikit-learn
torchvision